added this overview file
also finalizing a series of changes to make the structure of the projects, their dependencies, and build files parallel and systematic

in particular: changed all eclipse dependencies such that projects depend on jars in the deploy folder and not on other projects.
These dependencies are mediated through the deploy-wrapper project in the deploy folder, which everybody has to import now.

git-svn-id: https://svn.kwarc.info/repos/MMT/src@1305 09069c07-60fa-4895-8316-727242e252e4