Bugfixing QueryExtension and LMFDB extiension
Previously the QueryExtension class was called with an I() instance
instead of the subquery wrapped by it. This caused the
LMFDBQueryEvaluator problems, as it was expecting an un-wrapped query.
This commit addesses the problem by calling the QueryExtenstion class
with the unwrapped query.

Additionally, this commit makes some minor improvements and bugfixes
inside the LMFDB classes. In particular, queries passed to lmfdbquery()
now automatically resolve paginations. On top of this, calls to
LMFDB.fromPath can now be restricted to schema and db paths. This is
used to safeguard against racing conditions inside loading of the LMFDB
extension.