Implement an SCSCPServer component
This commit implements an SCSCPServer class that can be used to serve
functions to a client via SCSCP. It is a very basic implementation in
the sense that it does not integrate directly with MMT, but is a stand
alone part at the moment.

Additionally, this commit addresses a minor issue regarding parsing of
empty OMSTRs inside the SCSCPServer / SCSCPClient.