Update travis build script
Up until now, the .travis.yml only built the master branch. This made
sense for the earlier contribution model. With the new branches, all
branches should be built to check if a branch can be merged or not.

This commit removes the restriction from the travis build script and
allows all branches as well as pull requests to be built.