Add hostname support to the webserver
Previously the mmt webserver could only bind to all interfaces at once.
This can be a security risk, in particular when a machine has a
publically exposed interface (see for example KWARC/MathHub#243).

This commit fixes the issue by adding a new (optional) argument to the
ServerOn() command specificing the hostname to bind to. For example, one
can now start a server that only listens on localhost by typing `server
on 8080 localhost`. For backwards compatibility, the hostname argument
to the command is optional and defaults to 0.0.0.0.

In order to implement this argument some changes to the bundled tiscaf
library had to be made. In accordance with the LGPL license, the changes
have been noted and dated in the respective files.