Improvements to the webserver
This commit makes two improvements to the builtin webserver, one
intended towards MMT developers and one intended towards MMT users.

For MMT developers it is now easier to send a custom HTML Status Code
back to the user. This has been achieved by adding an optional argument
"status" of type "HStatus.Value" to all Server.*Response() methods.

For MMT users, the error response now tries to automatically pick
a content type for error messages. This is one of "text/plain",
"text/xml", "text/html" in this order with increasing priority. The
selecion is based on the "Accept:" header of the request. If no such
header is found, the server defaults to "text/html" (which was the case
until now).