Make NamespaceMap argument to Path.parse* optional
All of the Path.parse*() and related methods have a second argument of
type NamespaceMap. In many situations, no such NamespaceMap is available
and we simply use NamespaceMap.empty. This commit makes the second
argument to those methods optional, defaulting to NamespaceMap.empty.
Furthermore we now omit the argument in cases where just the default is
passed.