More aggressive CORS headers for Webserver
By default MMT only sent ```Access-Control-Allow-Origin``` headers when
an ```Origin:``` header was in the original request. In some situations,
this prevented some cross-domain scripts from working because they did
not send such request headers.

This commit fixes the issue by always sending CORS response headers,
even if no Origin Header is present. In case of an Origin being present,
it is used in the response, otherwise a wildcard ("*") is used.