Improvements to the MMT REPL
Up until now, the MMT shell has always been very basic. There was no
support for any kind of history and no kind of prompt. This commit
switches over to using the jLine library for a better shell and lays
foundations for future improvements.

In particular, support for navigating history with the arrow keys and a
prompt on the left hand side have been added. Furthermore, this commit
lays the basics for future improvements, like prompts on the right hand
side and tab completion support.