Refactor REPL functionality into REPLExtension
The new REPL functionality introduced in b5bf0d2 added a new external
dependency into the mmt-api project. According to @florian-rabe mmt-api
should remain dependency-free.

To work around this problem, while at the same time keeping the new
functionality, this commit introduces a new concept of REPLExtensions. A
REPLExtension provides a read-eval-print loop to MMT.

When MMT now starts a shell from the command line, the frontend will
request a list of all REPLExtensions from the ExtensionManager and use
the first one it finds. If no REPLExtension is loaded, it will
automatically load and activate StandardREPL, which is a
re-implementation of the old REPL behaviour.

The new jLine powered REPL (now named ExtendedREPL) has been refactored
into a seperate project, mmt-repl, which has an external dependency on
the jLine3 library.