Add a method to retrieve active log groups
In certain situations, for example when tab completing actions, it can
be useful to be aware of a list of active log groups. Up until now, it
was not possible to retrieve such a list outside of the mmt-api project.

This commit fixes the issue by modifying the ```Report``` class. The
list of active groups is stored inside of a ```mutable.Set``` which
should not be exposed directly.  Instead this commit adds a method
```active``` which retrieves a (read-only) list of active logging
groups.