Better dumb terminal support in ExtendedREPL
Previously, when the ExtendedREPL was launched in a 'dumb' terminal (one
without support for tabcompletion) all the advanced features were
quietly unused.

This commit changes the behaviour so that inside of a dumb terminal a
warning message is printed. Furthermore it allows users to prefix a
(partial) command with an '@' sign to have a list of suggestions printed
out the console.