Improve TabCompletion supoport in ExtendedREPL
Up until now TabCompletion support inside the ExtendedREPL was context
free (in the sense that it did not depend on the state of the
controller). This prevented it from making smart suggestions (for
example) for active log prefixes.

This commit makes the controller available to the ActionGrammar and uses
it for a few different commands. Furthermore it makes multiple
improvements to the CompletionEngine, allowing cleaner usage.