Add Sourceable trait and toSourceString methods
This commit adds a toSourceString() method to all objects inside of MMT.
is supposed to turn an object into a string that is equivalent to the
scala source code of the object.

Objects importing this method implement a new Sourceable trait. The
trait comes together with a type-safe application to all objects,
weather provided directly by MMT or not.