Cleanup magic web server paths
This commit cleans up and generalises the method for inlcuding MMT
webserver scripts.

Whenever MMT is run behind a proxy at a non-root directory (for example
in MathHub), it is not possible to properly include Resources from the
webserver because of the unknown root path.  This commit implements a
solution that rewrites requests of the form
"?do=mmt_resource&path=$resource" to the resource $resource.
Furthermore, it cleans up the existing mmt-url.js script which
automatically determines the root path for further requests on the
client side.