Better Error Handling for handle() and handleLine()
Previously the handle() and handleLine() controller methods did not
provide any useful error handling. The handle() command did not do any
kind of error handling, having ecxceptions bobble up to the caller. The
handleLine() command caught all errors, and did not provide any
machine-readable feedback on what went wrong.

In order to be able to write a proper shell that can understand the
result of the command, this commit instead introduces a new
`ActionResult` class which is returned from each handeled action.

This class indicates if the action was performed normally, if an error
occured during parsing of the action, or if an error occured during the
evaluation of the action. Furthermore, it has an `throwErrorIfAny()`
method that throws an exception if one occured during the execution of
the action.