sbt 1.x Compatibility (fixes #231)
Previously, the build.sbt file was not compatible with the sbt version
1.x, resulting in forced downgrades onto sbt 0.12.x and 0.13.x for
several users.

This commit introduces compatibility for sbt 1.x and cleans up the
build.sbt file. The compatibility is achieved by updating all used sbt
plugins to the most recent version.  The cleanup inlines a bunch of the
compilation code, and as a result results in fewer files in the project/
directory. As a result of this simplification, the `postProcessApi`
target is no longer neccessary.