Fix build.sbt regression
The cleaned up travis build process showed that two parts of build.sbt
were broken during recent commits. The API documentation fails to
build, and the `deployFull` task did not generate a fat mmt.jar.

This commit fixes the issues by introducing appropriate fixes to
`build.sbt`. To fix the first issue, the sbt-unidoc plugin is enabled
for the root sbt project instead of the mmt project. To fix the second
issue, we ensure that the `deployFull` task depends on the `deployFull`
task.