Structure the Travis CI Build Process
Previously the travis build process was a sequential list of commands
that was run. This was very unstructured, and a failing test provided
little information about what was going wrong.

This commit structures the travis build process by introducing different
stages. Build stages are a beta feature of Travis CI that allows to
specify a sequential list of stages, each containing a set of jobs that
is run in parallel.

We define five stages:
1. `build.sbt` that checks if the sbt build definition is valid;
2. `CodeCheck` that checks if the code compiles and also prints
scalastyle warnings;
3. `DeployCheck` that checks if the fatjar generation and API
documentation generation works;
4. `Test` that uses the tests defined in scala code; and
5. `Deploy` which auto-generates the API documentation and deploys it
to https://uniformal.github.io/apidoc/

Stages 1 - 4 are set to be run in all situations, stage 5 will only be
run on the master branch.