Replace sun.misc.BASE64* classes (fixes #260) (#264)
The mmt-odk project used the `sun.misc.BASE64Encoder` and `sun.misc.BASE64Decoder` classes, which are only available up to Java 8. This lead to compilation problems in Java 9.

This commit fixes the issue by replacing the classes with near-drop-in-replacements from `java.util.Base64`. This means that we have to drop support for Java <= 7 for our OpenMath implementation.