Fix sbt fatjar regression (closes #278)
In a previous commit, a regression was introduced that prevent the 'main
manifest attribute' to be set within jar generated by `sbt deploy` and
`sbt deployFull`.

The regression was caused by a scoping issue within sbt keys. This
commit fixes the regression by explicitly setting the `mainClass in
assembly` taskKey.