Fix apidoc source links
Previously, the source links generated by the apidoc target did not
properly point to the online version of the source code. Instead, they
prepended an absolute path of the system they were generated on.

This commit fixes the links by setting the sourcepath scalac option and
appropriately adopting the scaladoc options.