Fix pdflatex target
Due to a scala bug in the Process.cat function, the pdflatex target was
broken for files that required a pre- and post-amble. Previously, the
target concatinated the input files and passed them via STDIN to
pdflatex.

Instead, this commit works around the issue by creating a temporary
input file, which is then passed as a normal command line argument to
pdflatex.