Remove tiscaf and lfcatalog binary jar Dependencies
Previously, when tiscaf or lfcatalog were depended on as binary jars in
the deploy/ folder. To update them it was neccessary, to perform the
following steps exactly in this order:
* cleanly build tiscaf (using `sbt deployFull`)
* cleanly build lfcatalog (using `sbt deployFull` a
second time)
* cleanly rebuilding the mmt jar (`sbt deploy`)

This had two major disadvantages:
1) it was neccessary to rebuild the project 3 times to be sure that it
had been fully updated and
2) Several binary jars are under version control, resulting in code
being comitted twice and a large repository size.

The first one of these problems occured and went unnoticed until MathHub
attempted to start lfcatalog and complained.

This commit resolves the problem by updating build.sbt to not depend on
unmanaged tiscaf and lfcatalog jars. Instead it replaces them with
dependencies directly on the project.