Allow LMH to better handle non-archives
Previously, when atttempting to install anything that is not an archive,
LMH would crash. This would cause problems when e.g. a meta-inf archive
needed to be installed in a MathHub setting.

This commit splits the LMHHubEntry class into three possible
sub-classes, an LMHHubArchiveEntry, an LMHHubGroupEntry and an
LMHHubDirectoryEntry. These implement archives, "meta-inf" group
archives and non-semantic directories respectively.

Furthermore, we enforce that when an archive is installed, the
corresponding meta-inf repository is installed by lmh as well.