Added a new module mmt-got
The new module mmt-got provides some automated assistance for optimizing
theory graphs.

The GraphOptimizationTool is a new Extension.
It can look for certain possible optimizations in a theory.
It can look for possible optimizations in a collection of
theories.
It can look for possible optimizations in every theory of every present
archive.
Output can be converted to human readable xml.

Calling the optimizer is currently limited to an API.