Refactor MiTMServer
This commit refactors the MiTMServer. Instead of starting the SCSCP
Server by default, it now adds actions "scscp on [port [hostname]]" and
"scscp off" to control the server.

Furthermore this commit adds a new abstraction "MMTHandler" which allows
straightforward implementation of SCSCP Heads as OMA -> Term functions.
This abstraction is then used to re-implement the mitm eval head.