Add custom MMT_TODO Annotation
Previously, @deprecated annotation were used within the code to mark
parts of the code and note that they needed review and/or refactor.
This is different from the normal semantics assigned to this annotation,
and as such resulted in a lot of warnings that could not be easily
filtered or ignored.

This commit adds a new MMT_TODO annotation replacing the previous
deprecated annotations. This does not have any pre-defined semantics,
and is thus more suited.