Move to GitHub actions
This commit moves from Travis CI to GitHub Actions. GitHub Actions works
similar to Travis except that it is integrated directly with GitHub and
is easier to configure.

This commit removes now-defunct travis-specific code. It furthermore
configures GitHub Actions to automatically build an `mmt.jar` to every
commit and make it available for download.