Refactor MathML Presenters
This commit refactors the existing (Content and Presentation) MathML
Presenters to be able to handle parallel markup better.

In particular, the .toCML method is replaced by a new ContentMathML
presenter that carries context and IDs of subterms. The existing
MathMLPresenter is renamed to PresentationMathMLPresenter, to make the
difference clear.

This commit also prepares the legacy MWS harvester to output parallel
markup.